Organise files to packages
