+++ src/main/java/sana/Ui.java
@@ -1,5 +1,9 @@
+package sana;
+
 import java.util.LinkedList;
 import java.util.Scanner;
+
+import sana.task.*;
 
 /**
  * Ui represents the user interface object that deals with user interaction
@@ -38,18 +42,18 @@ public class Ui {
     }
 
     /**
-     * Prints the string representation of the task
+     * Prints the string representation of the sana.task
      *
-     * @param task  task to be printed
+     * @param task  sana.task to be printed
      */
     public void printTaskInList(Task task) {
         System.out.println(task);
     }
 
     /**
-     * Prints the entire task list
+     * Prints the entire sana.task list
      *
-     * @param taskList  task list to be printed
+     * @param taskList  sana.task list to be printed
      */
     public void printTaskList(LinkedList<Task> taskList) {
         int index = 1;
@@ -72,7 +76,7 @@ public class Ui {
     /**
      * Prints Sana's message when tasks are marked or unmarked
      *
-     * @param isMarked   if task is marked or unmarked
+     * @param isMarked   if sana.task is marked or unmarked
      */
     public void markText(boolean isMarked) {
         if (isMarked) {
@@ -99,14 +103,14 @@ public class Ui {
     }
 
     /**
-     * This method prints to system out Sana's add task text
+     * This method prints to system out Sana's add sana.task text
      */
     public void addTaskText() {
-        System.out.println("I've added your task!");
+        System.out.println("I've added your sana.task!");
     }
 
     /**
-     * This method prints to system out Sana's delete task text
+     * This method prints to system out Sana's delete sana.task text
      */
     public void deleteTaskText() {
         System.out.println("Yay! I'll take this out!");
